Extracted coherent classes to separate files
