+++ FormTemplateMethod/src/CapitalStrategies.ts
@@ -1,3 +1,6 @@
+import { Loan } from './Loans'
+import { RiskFactors } from './RiskFactors'
+
 abstract class CapitalStrategy {
   readonly MILLIS_PER_DAY: number = 8640000
   readonly DAYS_PER_YEAR: number = 365
@@ -15,15 +18,6 @@ abstract class CapitalStrategy {
   yearsTo(endDate: Date, loan: Loan): number {
     const beginDate: Date = (loan.getToday() == null ? loan.getStart() : loan.getToday())
     return ((endDate.getTime() - beginDate.getTime()) / this.MILLIS_PER_DAY / this.DAYS_PER_YEAR)
-  }
-}
-
-class RiskFactors {
-  forRating(riskRating: any): number {
-    return 2
-  }
-  static getFactors() {
-    return new RiskFactors()
   }
 }
 
@@ -78,53 +72,4 @@ class CapitalStrategyRevolver extends CapitalStrategy {
   }
 }
 
-class Loan {
-  unusedRiskAmount(): number {
-    return 10
-  }
-
-  getPayments(): LoanPayment[] {
-    return [new LoanPayment()]
-  }
-
-  getStart(): any {
-    return new Date("2020-09-01")
-  }
-
-  getToday(): Date {
-    return new Date("2020-09-29")
-  }
-
-  getExpiry(): Date {
-    return new Date("2021-09-01")
-  }
-
-  getRiskRating(): number {
-    return 1
-  }
-
-  outstandingRiskAmount(): number {
-    return 100  
-  }
-
-  public getCommitment(): number {
-    return 10
-  }
-
-  public getUnusedPercentage(): number {
-    return 10
-  }
-}
-
-class LoanPayment {
-  amount: number
-  date: Date
-
-  constructor() {
-    this.amount = 10
-    this.date = new Date("2020-09-15")
-  }
-
-}
-
-export { CapitalStrategyAdvisedLine, CapitalStrategyRevolver, CapitalStrategyTermLoan, Loan }
+export { CapitalStrategyAdvisedLine, CapitalStrategyRevolver, CapitalStrategyTermLoan }
