+++ FormTemplateMethod/src/RiskFactors.ts
@@ -1,0 +1,10 @@
+class RiskFactors {
+  forRating(riskRating: any): number {
+    return 2
+  }
+  static getFactors() {
+    return new RiskFactors()
+  }
+}
+
+export { RiskFactors }
